+++ app/models/post.rb
@@ -148,17 +148,20 @@ class Post < ActiveRecord::Base
     self.author.profile.nsfw?
   end
 
-  def self.find_by_guid_or_id_with_user(id, user=nil)
-    key = id.to_s.length <= 8 ? :id : :guid
-    post = if user
-             user.find_visible_shareable_by_id(Post, id, :key => key)
-           else
-             Post.where(key => id).includes(:author, :comments => :author).first
-           end
+  def self.find_public(id)
+    where(post_key(id) => id).includes(:author, comments: :author).first.tap do |post|
+      raise ActiveRecord::RecordNotFound.new("could not find a post with id #{id}") unless post
+      raise Diaspora::NonPublic unless post.public?
+    end
+  end
 
-    # is that a private post?
-    raise(Diaspora::NonPublic) unless user || post.try(:public?)
+  def self.find_non_public_by_guid_or_id_with_user(id, user)
+    user.find_visible_shareable_by_id(Post, id, key: post_key(id)).tap do |post|
+      raise ActiveRecord::RecordNotFound.new("could not find a post with id #{id}") unless post
+    end
+  end
 
-    post || raise(ActiveRecord::RecordNotFound.new("could not find a post with id #{id}"))
+  def self.post_key(id)
+    id.to_s.length <= 8 ? :id : :guid
   end
 end